==> 03_wkt_extraction.ipynb <==
#Extract wkt text from NMEP city shape files
# Warri, Minna, Osogbo and Zaria hvae errors in their wkt. Tried reprojecting crs, did not work 

import fiona
from shapely.geometry import shape
import geopandas as gpd
from shapely.ops import unary_union

# Warri
shapefile_path = 'C:/Users/USER/Urban Malaria Proj Dropbox/urban_malaria/data/nigeria/shapefiles/ShinyApp_shapefiles/Warri/Warri.shp'
gdf = gpd.read_file(shapefile_path)

combined_polygon = unary_union(gdf['geometry'])
combined_gdf = gpd.GeoDataFrame([1], geometry=[combined_polygon], crs= 4326)

print(combined_polygon)


##WARRI AGAIN
shapefile_path = 'C:/Users/USER/Urban Malaria Proj Dropbox/urban_malaria/data/nigeria/shapefiles/ShinyApp_shapefiles/Warri/Warri.shp'
gdf = gpd.read_file(shapefile_path)

gdf = gdf.to_crs(epsg=4326)
combined_polygon = unary_union(gdf['geometry'])
combined_gdf = gpd.GeoDataFrame([1], geometry=[combined_polygon], crs='EPSG:4326')

print(combined_polygon)

#Asaba
shapefile_path = 'C:/Users/USER/Urban Malaria Proj Dropbox/urban_malaria/data/nigeria/nigeria_shapefiles/shapefiles/ShinyApp_shapefiles/Asaba/Asaba.shp'
gdf = gpd.read_file(shapefile_path)

combined_polygon = unary_union(gdf['geometry'])
combined_gdf = gpd.GeoDataFrame([1], geometry=[combined_polygon], crs=gdf.crs)

print(combined_polygon)



# DELTA STATE
shapefile_path = 'C:/Users/USER/Urban Malaria Proj Dropbox/urban_malaria/data/nigeria/shapefiles/NMEP Net Distribution States Shapefiles for ADM3 - 13 States/Delta/Delta_Wards.shp'
gdf = gpd.read_file(shapefile_path)

combined_polygon = unary_union(gdf['geometry'])
combined_gdf = gpd.GeoDataFrame([1], geometry=[combined_polygon], crs=gdf.crs)

print(combined_polygon)


#Abeokuta
shapefile_path = 'C:/Users/USER/Urban Malaria Proj Dropbox/urban_malaria/data/nigeria/shapefiles/ShinyApp_shapefiles/Abeokuta/Abeokuta.shp'
gdf = gpd.read_file(shapefile_path)

combined_polygon = unary_union(gdf['geometry'])
combined_gdf = gpd.GeoDataFrame([1], geometry=[combined_polygon], crs=gdf.crs)

print(combined_polygon)

#Damaturu
shapefile_path = 'C:/Users/USER/Urban Malaria Proj Dropbox/urban_malaria/data/nigeria/shapefiles/ShinyApp_shapefiles/Damaturu/Damaturu.shp'
gdf = gpd.read_file(shapefile_path)

combined_polygon = unary_union(gdf['geometry'])
combined_gdf = gpd.GeoDataFrame([1], geometry=[combined_polygon], crs=gdf.crs)

print(combined_polygon)

#Dutse
shapefile_path = 'C:/Users/USER/Urban Malaria Proj Dropbox/urban_malaria/data/nigeria/shapefiles/ShinyApp_shapefiles/Dutse/Dutse.shp'
gdf = gpd.read_file(shapefile_path)

combined_polygon = unary_union(gdf['geometry'])
combined_gdf = gpd.GeoDataFrame([1], geometry=[combined_polygon], crs=gdf.crs)

print(combined_polygon)

#Gombe
shapefile_path = 'C:/Users/USER/Urban Malaria Proj Dropbox/urban_malaria/data/nigeria/shapefiles/ShinyApp_shapefiles/Gombe/Gombe.shp'
gdf = gpd.read_file(shapefile_path)

combined_polygon = unary_union(gdf['geometry'])
combined_gdf = gpd.GeoDataFrame([1], geometry=[combined_polygon], crs=gdf.crs)

print(combined_polygon)

#Ilorin
shapefile_path = 'C:/Users/USER/Urban Malaria Proj Dropbox/urban_malaria/data/nigeria/shapefiles/ShinyApp_shapefiles/Ilorin/Ilorin.shp'
gdf = gpd.read_file(shapefile_path)

combined_polygon = unary_union(gdf['geometry'])
combined_gdf = gpd.GeoDataFrame([1], geometry=[combined_polygon], crs=gdf.crs)

print(combined_polygon)

#Jalingo
shapefile_path = 'C:/Users/USER/Urban Malaria Proj Dropbox/urban_malaria/data/nigeria/shapefiles/ShinyApp_shapefiles/Jalingo/Jalingo.shp'
gdf = gpd.read_file(shapefile_path)

combined_polygon = unary_union(gdf['geometry'])
combined_gdf = gpd.GeoDataFrame([1], geometry=[combined_polygon], crs=gdf.crs)

print(combined_polygon)

#Kano
shapefile_path = 'C:/Users/USER/Urban Malaria Proj Dropbox/urban_malaria/data/nigeria/shapefiles/ShinyApp_shapefiles/Kano/Kano.shp'
gdf = gpd.read_file(shapefile_path)

combined_polygon = unary_union(gdf['geometry'])
combined_gdf = gpd.GeoDataFrame([1], geometry=[combined_polygon], crs=gdf.crs)

print(combined_polygon)

#Katsina
shapefile_path = 'C:/Users/USER/Urban Malaria Proj Dropbox/urban_malaria/data/nigeria/shapefiles/ShinyApp_shapefiles/Katsina/Katsina.shp'
gdf = gpd.read_file(shapefile_path)

combined_polygon = unary_union(gdf['geometry'])
combined_gdf = gpd.GeoDataFrame([1], geometry=[combined_polygon], crs=gdf.crs)

print(combined_polygon)

#Minna
shapefile_path = 'C:/Users/USER/Urban Malaria Proj Dropbox/urban_malaria/data/nigeria/shapefiles/ShinyApp_shapefiles/Minna/Minna.shp'
gdf = gpd.read_file(shapefile_path)

combined_polygon = unary_union(gdf['geometry'])
combined_gdf = gpd.GeoDataFrame([1], geometry=[combined_polygon], crs=gdf.crs)

print(combined_polygon)

#Osogbo
shapefile_path = 'C:/Users/USER/Urban Malaria Proj Dropbox/urban_malaria/data/nigeria/shapefiles/ShinyApp_shapefiles/Osogbo/Osogbo.shp'
gdf = gpd.read_file(shapefile_path)

combined_polygon = unary_union(gdf['geometry'])
combined_gdf = gpd.GeoDataFrame([1], geometry=[combined_polygon], crs=4326)

print(combined_polygon)

#Zaria
shapefile_path = 'C:/Users/USER/Urban Malaria Proj Dropbox/urban_malaria/data/nigeria/shapefiles/ShinyApp_shapefiles/Zaria/Zaria.shp'
gdf = gpd.read_file(shapefile_path)

combined_polygon = unary_union(gdf['geometry'])
combined_gdf = gpd.GeoDataFrame([1], geometry=[combined_polygon], crs= 4326)

print(combined_polygon)
